add firebase.js
